Stop trying to cleanup on Travis, since its sudoness

+++ test/zz-teardown.js
@@ -14,25 +14,33 @@ var rdes = path.resolve(__dirname, 'fixtures', '.registry_design')
 var udes = path.resolve(__dirname, 'fixtures', '._users_design')
 var dotDelete = path.resolve(__dirname, 'fixtures', '.delete')
 
-test('cleanup', function (t) {
-  try {
-    var pid = fs.readFileSync(pidfile)
-  } catch (er) {}
-
-  if (pid) {
-    try { process.kill(pid) } catch (er) {
-      // ok if already killed
-      t.equal(er.code, 'ESRCH')
-    }
-  }
-  var files = [ pidfile, repl, log, _users, db, rdes, dotDelete ]
-  if (!process.env.TRAVIS) {
-      files.push(udes);
-  }
-  files.forEach(function(file) {
-    rimraf.sync(file)
+if (process.env.TRAVIS) {
+  test('nocleanup', function (t) {
+    t.pass('No cleanup on travis')
+    t.end()
   })
+}
+else {
+  test('cleanup', function (t) {
+    try {
+      var pid = fs.readFileSync(pidfile)
+    } catch (er) {}
 
-  t.pass('couch is no more')
-  t.end()
-})
+    if (pid) {
+      try { process.kill(pid) } catch (er) {
+        // ok if already killed
+        t.equal(er.code, 'ESRCH')
+      }
+    }
+    var files = [ pidfile, repl, log, _users, db, rdes, dotDelete ]
+    if (!process.env.TRAVIS) {
+        files.push(udes);
+    }
+    files.forEach(function(file) {
+      rimraf.sync(file)
+    })
+
+    t.pass('couch is no more')
+    t.end()
+  })
+}
